requirements.txt
django>=1.11,<1.12
django-debug-toolbar==1.8
gunicorn==19.4.5
psycopg2==2.8.3
whitenoise==3.3.1
djangorestframework==3.9.4
scikit-learn==0.21.2
